Make RecordStoreCompiledCondition a separate java class for serialization
